update: implemented reset state action case for createGameMenu
This is to allow the CreateGameMenu container to reset the redux state for createGameMenu when the submit button is clicked and all checks pass. This will prevent users from going back the CreateGameMenu page in the redux history and it showing default Ui with nothing filled in. When in reality the state would still have been set from any previous submissions.