ver 0.1.31a
Minor changes and re-arrangements in code. Still need to definitively decide how / where to put checking for making socket request. Options are: 1. Do it only client side on each socket func leading to the server 'trusting' what it is told to do by the client, 2. Do it both client and server side requiring the sending of extra data to ensure correct/necessary operations are done, 3. Only do it server side but potentially receive erroneous or useless request from client. Due to using socket for the game part it may be worth doing client side checks and server and then updating the client state instantly if the check passes client side which removes the lag waiting for the server to tell the client to update.