update: ver 1.0.0
Application is in a fully working state with minor known bugs. Future additions to be made are implementing a "how to" or "rules" button that provides information on how to play the game and animations for the word blocks, buttons etc. Other features may be created but are not currently planned.